# docs/swagger.json
{
    "components": {"schemas":{"handlers.APIInfo":{"description":"General Info about the API, author name, how to contact, etc.","properties":{"author":{"description":"Author name.","example":"Kareem Ebrahim","type":"string"},"contact":{"description":"How to contact the author.","example":"kareemmahlees@gmail.com","type":"string"},"repo":{"description":"Git repository name for contributions.","example":"https://github.com/kareemmahlees/meta-x","type":"string"},"year":{"description":"Year of launch","example":2024,"type":"integer"}},"type":"object"},"models.AddModifyColumnPayload":{"description":"Adding or Modifying a column payload.","properties":{"column_name":{"description":"Column Name","type":"string"},"type":{"description":"New type","example":"Int","type":"string"}},"required":["column_name","type"],"type":"object"},"models.BadRequestError":{"description":"Request payload failed validation.","properties":{"code":{"description":"HTTP code.","example":422,"type":"integer"},"message":{"description":"Error cause.","example":"Payload Failed Validation"}},"type":"object"},"models.CreatePgMySqlDBPayload":{"properties":{"name":{"description":"Database name.","example":"Users","type":"string"}},"required":["name"],"type":"object"},"models.CreateTablePayload":{"description":"Data about the column to create.","properties":{"column_name":{"description":"Name of the column.","type":"string"},"default":{"description":"Default value of the column."},"nullable":{"description":"If the column accepts null values or not."},"type":{"description":"Data type of the column.","type":"string"},"unique":{"description":"Wether to add a unique constraint on the column."}},"required":["column_name","type"],"type":"object"},"models.CreateTableResp":{"description":"Table Created Successfully.","properties":{"created":{"description":"Created table name.","type":"string"}},"type":"object"},"models.DeleteColumnPayload":{"description":"Column to delete data.","properties":{"column_name":{"description":"Column Name","type":"string"}},"required":["column_name"],"type":"object"},"models.ErrResp":{"description":"Operation Failed. Maybe invalid payload or internal server error.","properties":{"code":{"description":"HTTP code.","example":400,"type":"integer"},"message":{"description":"Error cause.","example":"Malformed payload"}},"type":"object"},"models.InternalServerError":{"description":"Something wrong happened on the server.","properties":{"code":{"description":"HTTP code.","example":500,"type":"integer"},"message":{"description":"Error cause.","example":"Something went wrong"}},"type":"object"},"models.ListDatabasesResp":{"description":"A list of all available databases.","properties":{"databases":{"example":["test","prod","main"],"items":{"type":"string"},"type":"array","uniqueItems":false}},"type":"object"},"models.ListTablesResp":{"description":"List of tables.","properties":{"tables":{"example":["table1","table2","table3"],"items":{"type":"string"},"type":"array","uniqueItems":false}},"type":"object"},"models.SuccessResp":{"description":"Operation Succeeded.","properties":{"success":{"type":"boolean"}},"type":"object"},"models.TableColumnInfo":{"description":"Info about a table column.","properties":{"default":{"description":"Default value of the column.","example":"123"},"key":{"description":"Constraint name on the column.","example":"PRI"},"name":{"description":"Name of the column.","example":"email","type":"string"},"nullable":{"description":"If the column accepts null values or not.","example":"true","type":"string"},"type":{"description":"Data type of the column.","example":"varchar","type":"string"}},"type":"object"},"models.UnprocessableEntityError":{"description":"Failed to parse payload.","properties":{"code":{"description":"HTTP code.","example":422,"type":"integer"},"message":{"description":"Error cause.","example":"Payload Parsing Failed"}},"type":"object"}}},
    "info": {"contact":{"email":"kareemmahlees@gmail.com","name":"Kareem Ebrahim"},"description":"A RESTFull and GraphQL API to supercharge your database","title":"MetaX","version":"0.1.1"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/":{"get":{"description":"Get general info about the API","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/handlers.APIInfo"}}},"description":"OK"}},"summary":"Get API Info","tags":["General"]}},"/database":{"get":{"description":"Get all the available databases.","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.ListDatabasesResp"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"List Databases","tags":["Database"]},"post":{"description":"Creates a new database with the specified table schema.\n\n**NOTE**: SQLite is Unsupported.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.CreatePgMySqlDBPayload"}}},"description":"Database Info","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.SuccessResp"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.ErrResp"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Create a new Database","tags":["Database"]}},"/table":{"get":{"description":"Get a list of the available tables in the database.","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.ListTablesResp"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"List tables.","tags":["Table"]}},"/table/{table_name}":{"delete":{"description":"Delete the given table.","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.SuccessResp"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Delete Table","tags":["Table"]},"post":{"description":"Creates a new table with the specified columns.","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/models.CreateTablePayload"},"type":"array"}}},"description":"Table Data","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.CreateTableResp"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.BadRequestError"}}},"description":"Bad Request"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.UnprocessableEntityError"}}},"description":"Unprocessable Entity"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Create a Table.","tags":["Table"]}},"/table/{table_name}/column/add":{"post":{"description":"Adds a column with the provided data to the given table.","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.AddModifyColumnPayload"}}},"description":"Column Data","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.SuccessResp"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.BadRequestError"}}},"description":"Bad Request"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.UnprocessableEntityError"}}},"description":"Unprocessable Entity"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Add a column","tags":["Table"]}},"/table/{table_name}/column/delete":{"delete":{"description":"Delete table column","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.DeleteColumnPayload"}}},"description":"Column Name","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.SuccessResp"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.BadRequestError"}}},"description":"Bad Request"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.UnprocessableEntityError"}}},"description":"Unprocessable Entity"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Delete Column","tags":["Table"]}},"/table/{table_name}/column/modify":{"put":{"description":"Update table column properties, **Only supports updating the column type for now**.","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.AddModifyColumnPayload"}}},"description":"Column Data","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.SuccessResp"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.BadRequestError"}}},"description":"Bad Request"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.UnprocessableEntityError"}}},"description":"Unprocessable Entity"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Update Column","tags":["Table"]}},"/table/{table_name}/describe":{"get":{"description":"Get detailed info about a table's fields.","parameters":[{"description":"Table Name","in":"path","name":"table_name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/models.TableColumnInfo"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/models.InternalServerError"}}},"description":"Internal Server Error"}},"summary":"Get table info","tags":["Table"]}}},
    "openapi": "3.1.0",
    "tags": [
        {"description":"general info about the API.","name":"General"},
        {"description":"Database level related operations.","name":"Database"},
        {"description":"Table Level related operations.","name":"Table"}
    ],
    "servers": [
        {"description":"Home town of Meta-X","url":"http://localhost:5522"}
    ]
}